+++ ext_tables.php
@@ -1,6 +1,0 @@
-<?php
-if (!defined('TYPO3_MODE')) {
-	die ('Access denied.');
-}
-
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile($_EXTKEY, 'Configuration/TypoScript', 'pdfviewhelpers');